% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_num_if_num}
\alias{as_num_if_num}
\title{convert something to numeric if it can be}
\usage{
as_num_if_num(value)
}
\arguments{
\item{value}{the value to attempt to convert}
}
\description{
convert something to numeric if it can be
}
