% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_tsv_type}
\alias{assign_tsv_type}
\title{Assign data types}
\usage{
assign_tsv_type(data)
}
\arguments{
\item{data}{the column}
}
\description{
This is necessary for proper importing into ecotaxa
this is written to be used with sapply
}
