% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processing.R
\name{batch_dup_deleter}
\alias{batch_dup_deleter}
\title{Batch Duplicate Deleter}
\usage{
batch_dup_deleter(root_dir, write_plots = TRUE, ...)
}
\arguments{
\item{root_dir}{the root directory. Will go into _work if present}

\item{write_plots}{boolean to write plots to _results directory}

\item{...}{additional agruments to pass to flag_duplicates}
}
\description{
This function will iterate through a directory of runs
and implement duplicate deleter
}
\author{
Alex Barth
}
