% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processing.R
\name{batch_flowprocess}
\alias{batch_flowprocess}
\title{Batch Establish Flowprocess}
\usage{
batch_flowprocess(root_dir, ...)
}
\arguments{
\item{root_dir}{the root directory. Will go into _work if present}

\item{...}{additional agruments to pass to establish_flowprocess}
}
\description{
This function will iterate through a directory of runs
and structure them to be flowprocess files
}
