% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processing.R
\name{batch_trim}
\alias{batch_trim}
\title{Batch Trim}
\usage{
batch_trim(root_dir, min_esd)
}
\arguments{
\item{root_dir}{Root directory preferably the project level (above _work)}

\item{min_esd}{the minimum esd}
}
\description{
Batch Trim
}
