% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowprocess_file_formatter.R
\name{construct_etx_tsv}
\alias{construct_etx_tsv}
\title{Construct ecotaxa_tsv file}
\usage{
construct_etx_tsv(
  root_dir,
  raw_data_file,
  raw_summary_file,
  image_file_type = ".png"
)
}
\arguments{
\item{root_dir}{the root directory}

\item{raw_data_file}{the raw data as a data frame}

\item{raw_summary_file}{the raw summary as a dataframe}

\item{image_file_type}{the file type of exported images}
}
\description{
This takes in the summary data and the data file to construct
an ecotaxa tsv style file
}
