% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_deleter.R
\name{density_space_thresholder}
\alias{density_space_thresholder}
\title{Kernel Density Thresholding}
\usage{
density_space_thresholder(metadata, grid_size = 100, threshold_multiplier = 2)
}
\arguments{
\item{metadata}{a metadata file from a tsv function}

\item{grid_size}{the resolution of grid-points to use}

\item{threshold_multiplier}{number of sd away from mean density}
}
\description{
This is the first step to identify spatial overlap
}
