% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_deleter.R
\name{duplicate_deleter}
\alias{duplicate_deleter}
\title{Duplicate Delete}
\usage{
duplicate_deleter(root_dir, write_plots = TRUE, ...)
}
\arguments{
\item{root_dir}{the root director of your project}

\item{write_plots}{function to write plots}

\item{...}{arguments to pass on to flag_duplicates}
}
\description{
This function calls two_step_flag to get index of points
to delete but the main actions here are to
}
\author{
Alex Barth
}
