% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowprocess_file_formatter.R
\name{establish_flowprocess}
\alias{establish_flowprocess}
\title{Establish Folder structure for FlowProcess}
\usage{
establish_flowprocess(
  root_dir,
  meta_data = find_meta_files(root_dir),
  clean_files = T,
  image_file_type = ".png"
)
}
\arguments{
\item{root_dir}{The root directory}

\item{meta_data}{the metadata file}

\item{clean_files}{Boolean to remove original data/summary files}

\item{image_file_type}{the file format, default to '.png'}
}
\description{
Establish Folder structure for FlowProcess
}
