% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_deleter.R
\name{feature_cluster}
\alias{feature_cluster}
\title{feature clustering with HDBSCAN}
\usage{
feature_cluster(
  data,
  minPts = 5,
  features = c("object_esd", "object_abd", "object_feret", "object_blue", "object_red",
    "object_green", "object_circle_fit", "object_circularity", "object_circularity_hu",
    "object_compactness", "object_convex_perim", "object_convexivity",
    "object_edge_gradient", "object_elongation", "object_feret_max", "object_feret_min",
    "object_fiber_curl", "object_fiber_straight", "object_length", "object_width",
    "acq_source_img")
)
}
\arguments{
\item{data}{the data-preferably pre filtered on proximity}

\item{minPts}{The minimum number of points}

\item{features}{the features to cluster on}
}
\description{
feature clustering with HDBSCAN
}
