% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowprocess_file_formatter.R
\name{find_meta_files}
\alias{find_meta_files}
\title{Automatically detect the needed meta data in your root_dir}
\usage{
find_meta_files(root_dir)
}
\arguments{
\item{root_dir}{The Root directory}
}
\description{
Automatically detect the needed meta data in your root_dir
}
