% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_deleter.R
\name{find_read_tsv}
\alias{find_read_tsv}
\title{Find and read etx tsv}
\usage{
find_read_tsv(root_dir)
}
\arguments{
\item{root_dir}{root directory}
}
\description{
Works in a flowprocess-formatted folder
}
