% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_deleter.R
\name{flag_duplicates}
\alias{flag_duplicates}
\title{Pipeline for two-step flagging}
\usage{
flag_duplicates(
  metadata,
  grid_size = 100,
  threshold_multiplier = 2,
  minPts = 5,
  features = c("object_esd", "object_abd", "object_feret", "object_blue", "object_red",
    "object_green", "object_circle_fit", "object_circularity", "object_circularity_hu",
    "object_compactness", "object_convex_perim", "object_convexivity",
    "object_edge_gradient", "object_elongation", "object_feret_max", "object_feret_min",
    "object_fiber_curl", "object_fiber_straight", "object_length", "object_width",
    "acq_source_img")
)
}
\arguments{
\item{metadata}{the metadata frame}

\item{grid_size}{the resolution of grid-points to use}

\item{threshold_multiplier}{number of sd away from mean density}

\item{minPts}{The minimum number of points}

\item{features}{the features to cluster on}
}
\description{
This will feed the meta through first the space detector
then through the feature clustering cut stage
}
\author{
Alex Barth
}
