% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{no_leaders_no_trailers}
\alias{no_leaders_no_trailers}
\title{Drop leading and trailing spaces}
\usage{
no_leaders_no_trailers(value)
}
\arguments{
\item{value}{character value to drop}
}
\description{
Drop leading and trailing spaces
}
