% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_deleter.R
\name{plot_obs}
\alias{plot_obs}
\title{Plot the distribution of observations}
\usage{
plot_obs(metadata, dup_index = NULL)
}
\arguments{
\item{metadata}{the tsv file}

\item{dup_index}{duplicated points}
}
\description{
Plot the distribution of observations
}
\author{
Alex Barth
}
