% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowprocess_file_formatter.R
\name{read_data_file}
\alias{read_data_file}
\title{Read the raw data file}
\usage{
read_data_file(root_dir, data_file_name)
}
\arguments{
\item{root_dir}{the root directory}

\item{data_file_name}{the data file name}
}
\description{
This is an internal function to read in
}
