% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_etx_tsv}
\alias{read_etx_tsv}
\title{A short-cut way to read etx files}
\usage{
read_etx_tsv(path)
}
\arguments{
\item{path}{the pathway to the file}
}
\description{
Feeds to read.table
}
\author{
Alex Barth
}
