% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowprocess_file_formatter.R
\name{read_summary_file}
\alias{read_summary_file}
\title{Read the summary file}
\usage{
read_summary_file(root_dir, summary_file_name)
}
\arguments{
\item{root_dir}{The root file directory}

\item{summary_file_name}{The summary file name}
}
\description{
Because the summary file is whacky - we will pull it into a list
}
