% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_to_esd.R
\name{trim_to_esd}
\alias{trim_to_esd}
\title{Trim folders based on column}
\usage{
trim_to_esd(root_dir, min_esd)
}
\arguments{
\item{root_dir}{the folder of a project - needs a tsv}

\item{min_esd}{the minimum est to trim to}
}
\description{
If the flowcam autoimaging keeps toomany object of certain sizes,
you can delete the images here
}
